Initial prototype for fetching subjects by compound
This prototype allows the user to query tranSMART for the existence of
certain compoundIDs linked to concepts in the study tree. It then
returns the subjectIDs.

This first implementation circumvents the proper way of fetching data
from tranSMART by connecting directly to the database for fetching the
metadata-tags. TranSMART's RESTful API will be extended to allow
fetching of metadata tags, after which the RClient needs to use the
proper API calls instead of the direct sql query from the database.