Improved column selection for getConcept function
Now works as the other functions by defining columns to be included
instead of columns to be excluded. New version of RESTful hAPI as added
some janitorial columns to the call return. Instead of excluding these,
and likely other new ones in the near future, it is easer to define the
ones to be included.